+++ src/slskd/Integrations/Pushbullet/PushbulletService.cs
@@ -29,7 +29,7 @@ namespace slskd.Integrations.Pushbullet
     using System.Threading.Tasks;
     using Microsoft.Extensions.Caching.Memory;
     using Microsoft.Extensions.Logging;
-    using static slskd.Options.IntegrationsOptions;
+    using static slskd.Options.IntegrationOptions;
 
     /// <summary>
     ///     Pushbullet integration service.
@@ -60,7 +60,7 @@ namespace slskd.Integrations.Pushbullet
         private IHttpClientFactory HttpClientFactory { get; }
         private ILogger<PushbulletService> Log { get; }
         private IOptionsMonitor<Options> OptionsMonitor { get; }
-        private PushbulletOptions PushbulletOptions => OptionsMonitor.CurrentValue.Integrations.Pushbullet;
+        private PushbulletOptions PushbulletOptions => OptionsMonitor.CurrentValue.Integration.Pushbullet;
         private IMemoryCache RecentlySent { get; }
 
         /// <summary>